feat: make secrets installation opt-in

+++ setup/30-secrets.sh
@@ -12,16 +12,18 @@ create_1password_session() {
     fi
 }
 
-if [ ! -f "$HOME/.ssh/id_rsa" ] ; then
-    create_1password_session
-    mkdir -p "$HOME/.ssh"
-    /tmp/op get document "SSH Key" --vault "Device Setup" --output "$HOME/.ssh/id_rsa"
-fi
+if [ "$INSTALL_SECRETS" != "false" ] ; then
+    if [ ! -f "$HOME/.ssh/id_rsa" ] ; then
+        create_1password_session
+        mkdir -p "$HOME/.ssh"
+        /tmp/op get document "SSH Key" --vault "Device Setup" --output "$HOME/.ssh/id_rsa"
+    fi
 
-email=$(cat /tmp/email)
-if ! gpg --list-keys "$email" >/dev/null ; then
-    create_1password_session
-    /tmp/op get document "GPG Key" --vault "Device Setup" --output /tmp/gpg-key
-    gpg --import /tmp/gpg-key
-    rm /tmp/gpg-key
+    email=$(cat /tmp/email)
+    if ! gpg --list-keys "$email" >/dev/null ; then
+        create_1password_session
+        /tmp/op get document "GPG Key" --vault "Device Setup" --output /tmp/gpg-key
+        gpg --import /tmp/gpg-key
+        rm /tmp/gpg-key
+    fi
 fi
